+++ mod/Mods/services/DATIS/Options/optionsData.lua
@@ -1,12 +1,4 @@
 cdata = {
-  DCSATIS = _("DCS ATIS USING AMAZON POLLY AND GOOGLE TEXT TO SPEECH"),
-  DATIS_DEFAULT_VOICE = _("Default Voice:"),
-  DATIS_CLOUD_ACCESS_KEY = _("Google Cloud Access Key:"),
-  DATIS_AWS_ACCESS_KEY = _("Amazon Access Key:"),
-  DATIS_AWS_PRIVATE_KEY = _("Amazon Private Key:"),
-  DATIS_AWS_REGION = _("Amazon Region:"),
-  DATIS_AWS_REGION_TEXT = ("See available regions at https://docs.aws.amazon.com/general/latest/gr/rande.html#endpoint-tables."),
-  DATIS_SRS_SERVER_PORT = _("SRS Server Port (5002):"),
-  DATIS_DEBUG_LOGGING = _("Enable Debug Logging"),
-  DATIS_DEBUG_LOGGING_HELP = _("Add additional log output Saved Games\\Logs\\DATIS.log.")
+  DCSATIS = _("DATIS - DCS ATIS"),
+  DATIS_CONFIG_INFO = _("The config has been moved into `Saved Games/DCS.openbeta/Config/DATIS.json`")
 }